requirements.txt
keyboard==0.13.5
mss==6.1.0
numpy==1.22.0
opencv-python==4.5.5.62
Pillow==9.0.0
pynput==1.7.6
